requirements.txt
docker-py
gitpython
pyyaml
